fix one more portability problem about printf format string
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5052 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ regress/lib/libgfarm/gfarm/gfs_getxattr_cached/size0.c
@@ -87,7 +87,7 @@ main(int argc, char **argv)
 		    path, xattr, (int)sz0, (int)sz1, (int)sz2);
 		return (EXIT_FAILURE);
 	}
-	printf("path %s, xattr %s: size %d\n", path, xattr, sz0);
+	printf("path %s, xattr %s: size %d\n", path, xattr, (int)sz0);
 
 	e = gfarm_terminate();
 	if (e != GFARM_ERR_NO_ERROR) {
